Intent: decouple glyph runtime helper from validation Operators involved: [Coherence] Affected invariants: [] Key changes: - Move last_glyph helper to tnfr.glyph_runtime without validation imports - Import validate_window directly in glyph_history and update references - Adjust consumers, stubs, and validators to use the new glyph_runtime path Expected risks/dissonances: Minimal risk of import regressions; validated via targeted tests Metrics: C(t), Si, νf, phase expected unchanged Equivalence map: last_glyph => tnfr.glyph_runtime.last_glyph
