+++ src/tnfr/glyph_history.py
@@ -8,19 +8,10 @@ from itertools import islice
 from typing import Any, cast
 
 from .constants import get_param, normalise_state_token
-from typing import TYPE_CHECKING
-
+from .glyph_runtime import last_glyph
 from .types import TNFRGraph
 from .utils import ensure_collection, get_logger
-
-if TYPE_CHECKING:  # pragma: no cover - typing convenience
-    from .validation import validate_window as _validate_window
-
-
-def _validate_window(window: int, *, positive: bool = False) -> int:
-    from .validation import validate_window as _validate_window
-
-    return _validate_window(window, positive=positive)
+from .validation import validate_window
 
 logger = get_logger(__name__)
 
@@ -31,7 +22,6 @@ __all__ = (
     "ensure_history",
     "current_step_idx",
     "append_metric",
-    "last_glyph",
     "count_glyphs",
 )
 
@@ -41,7 +31,7 @@ def _ensure_history(
 ) -> tuple[int, deque[str] | None]:
     """Validate ``window`` and ensure ``nd['glyph_history']`` deque."""
 
-    v_window = _validate_window(window)
+    v_window = validate_window(window)
     if v_window == 0 and not create_zero:
         return v_window, None
     hist = nd.setdefault("glyph_history", deque(maxlen=v_window))
@@ -274,14 +264,6 @@ def append_metric(hist: MutableMapping[str, list[Any]], key: str, value: Any) ->
         return
 
     hist.setdefault(key, []).append(value)
-
-
-def last_glyph(nd: Mapping[str, Any]) -> str | None:
-    """Return the most recent glyph for node or ``None``."""
-    hist = nd.get("glyph_history")
-    return hist[-1] if hist else None
-
-
 def count_glyphs(
     G: TNFRGraph, window: int | None = None, *, last_only: bool = False
 ) -> Counter[str]:
@@ -293,7 +275,7 @@ def count_glyphs(
     """
 
     if window is not None:
-        window = _validate_window(window)
+        window = validate_window(window)
         if window == 0:
             return Counter()
