+++ src/tnfr/glyph_history.pyi
@@ -30,7 +30,6 @@ def current_step_idx(G: TNFRGraph | Mapping[str, Any]) -> int: ...
 def append_metric(
     hist: MutableMapping[str, list[Any]], key: str, value: Any
 ) -> None: ...
-def last_glyph(nd: Mapping[str, Any]) -> str | None: ...
 def count_glyphs(
     G: TNFRGraph, window: int | None = ..., *, last_only: bool = ...
 ) -> Counter[str]: ...
