+++ src/tnfr/glyph_runtime.pyi
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from collections.abc import Mapping
+from typing import Any
+
+__all__: tuple[str, ...]
+
+def last_glyph(nd: Mapping[str, Any]) -> str | None: ...
+
